agent/requirements.txt
openai
python-docx
google-api-python-client
google-auth
python-dotenv
